Moved more pipeline scanning functions to the pipelines pkg
